+++ stubs/functions.php
@@ -5,7 +5,7 @@
  * IDE autocompletion stubs. Do not include this file at runtime.
  *
  * @package JsonQ
- * @version 0.4.0
+ * @version 0.4.1
  * @license The PHP License, version 3.01
  */